chore: improve tools description (#518)
* chore: improve tools description

* fix: missing punctuation

* chore: update en/us version

* chore: create chinese version of releasearchive.md

* fix: revert commit

I apologize for my mistake that I added the commit to the wrong branch.

* Chore: Remove v2rayNG

Remove v2rayNG as it is currently not using v2ray-core any more.

* Chore: Mark v2rayU as not updating

V2rayU has already set the branch "rm" as the main branch, which means that the developer is not maintaining the repository any more.

* Chore: Update English version

* Chore: fix type error

* fix: revert error

* Chore: Mark nekoray as stopped maintaining

* Chore: remove v2rayU

Not using v2ray-core.

* Chore: Update en/us version